+++ src/main/java/calculator/view/OutputView.java
@@ -1,7 +1,7 @@
 package calculator.view;
 
 public class OutputView {
-    public void printOutput(int result) {
+    public void printOutput(long result) {
         System.out.print("결과 : " + result);
     }
 }
